+++ NEWS.md
@@ -1,3 +1,8 @@
+# RstoxData v2.1.5-9001  (2025-04-28)
+* Removed EchoType from ICESAcousticData, which should only appear if the present in the input file.
+* Added support for R 4.5 binary.
+
+
 # RstoxData v2.1.4  (2025-03-31)
 * Final release for StoX 4.1.3.
 
@@ -39,7 +44,7 @@ RstoxData v2.1.3  (2025-03-21)
 # RstoxData v2.1.1_9001  (2024-11-25)
 * Removed rscript_args from mapplyOnCores(), since this caused the Renviron to be ignored in the sub processes, resulting in errors with finding RstoxFramework in the case of multiple libraries on Linux and macOS.
 * Fixed the 29 cm bug again, as it was not properly fixed in RstoxData v2.1.0. Refactored how precision is set both when reading data and in the ICESBiotic() and ICESAcoustic().
-* Introduced the EchoType as a column in the Data table of ICESBiotic() (and WriteICESBiotic()).
+* Introduced the EchoType as a column in the Data table of ICESAcoustic() (and WriteICESAcoustic()).
 
 
 # RstoxData v2.1.0  (2024-11-04)